Fixed feature 2

+++ src/main/java/com/apap/tugas1/repository/JabatanPegawaiDb.java
@@ -1,7 +1,6 @@
 package com.apap.tugas1.repository;
 
 import java.math.BigInteger;
-import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -13,6 +12,6 @@ import com.apap.tugas1.model.JabatanPegawaiModel;
  *  @author Priscilla Tiffany
  */
 @Repository
-public interface JabatanPegawaiDb extends JpaRepository<JabatanPegawaiModel, Long>{
+public interface JabatanPegawaiDb extends JpaRepository<JabatanPegawaiModel, BigInteger>{
 //	List<JabatanPegawaiModel> findByIdJabatan(BigInteger id);
 }
